Add a `MakeWriter` param to HierarchicalLayer
