+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tracing-tree"
-version = "0.1.2"
+version = "0.1.3"
 authors = ["David Barsky <me@davidbarsky.com>", "Nathan Whitaker"]
 edition = "2018"
 license = "MIT OR Apache-2.0"
@@ -10,7 +10,7 @@ description = "a tracing layer that represents prints a heirarchal tree of spans
 
 [dependencies]
 tracing = "0.1"
-tracing-subscriber = "0.2"
+tracing-subscriber = { version = "0.2", default-features = false, features = ["registry", "fmt"] }
 quanta = "0.3.1"
 termcolor = "1.0.5"
 ansi_term = "0.12.1"
