+++ src/lib.rs
@@ -1,109 +1,32 @@
+pub(crate) mod format;
+
 use ansi_term::{Color, Style};
 use chrono::{DateTime, Local};
-use std::sync::Mutex;
+use format::{Buffers, ColorLevel, Config, FmtEvent};
 use std::{
+    default::Default,
     fmt::{self, Write as _},
     io,
+    sync::Mutex,
 };
-
 use tracing::{
     field::{Field, Visit},
     span::{Attributes, Id},
-    Event, Level, Subscriber,
+    Event, Subscriber,
 };
 use tracing_subscriber::{
+    fmt::MakeWriter,
     layer::{Context, Layer},
     registry::LookupSpan,
 };
 
-const LINE_VERT: &str = "│";
-const LINE_HORIZ: &str = "─";
-const LINE_BRANCH: &str = "├";
-
-#[derive(Debug)]
-struct Config {
-    ansi: bool,
-    indent_lines: bool,
-    indent_amount: usize,
-}
-
-impl Config {
-    fn with_ansi(self, ansi: bool) -> Self {
-        Self { ansi, ..self }
-    }
-    fn with_indent_lines(self, indent_lines: bool) -> Self {
-        Self {
-            indent_lines,
-            ..self
-        }
-    }
-}
-
-impl Default for Config {
-    fn default() -> Self {
-        Self {
-            ansi: true,
-            indent_lines: false,
-            indent_amount: 2,
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct HierarchicalLayer {
-    stdout: io::Stdout,
-    bufs: Mutex<Buffers>,
-    config: Config,
-}
-
-#[derive(Debug)]
-struct Buffers {
-    pub current_buf: String,
-    pub indent_buf: String,
-}
-
-impl Buffers {
-    fn new() -> Self {
-        Self {
-            current_buf: String::new(),
-            indent_buf: String::new(),
-        }
-    }
-
-    fn flush_current_buf(&mut self, mut writer: impl io::Write) {
-        write!(writer, "{}", &self.current_buf).unwrap();
-        self.current_buf.clear();
-    }
-
-    fn flush_indent_buf(&mut self) {
-        self.current_buf.push_str(&self.indent_buf);
-        self.indent_buf.clear();
-    }
-
-    fn indent_current(&mut self, indent: usize, config: &Config) {
-        indent_block(
-            &mut self.current_buf,
-            &mut self.indent_buf,
-            indent,
-            config.indent_amount,
-            config.indent_lines,
-        );
-        self.current_buf.clear();
-    }
-}
-
-struct Data {
+pub struct Data {
     start: DateTime<Local>,
     kvs: Vec<(&'static str, String)>,
 }
 
-struct FmtEvent<'a> {
-    bufs: &'a mut Buffers,
-    comma: bool,
-}
-
 impl Data {
-    fn new(attrs: &tracing::span::Attributes<'_>) -> Self {
+    pub fn new(attrs: &tracing::span::Attributes<'_>) -> Self {
         let mut span = Self {
             start: Local::now(),
             kvs: Vec::new(),
@@ -119,138 +42,61 @@ impl Visit for Data {
     }
 }
 
-impl<'a> Visit for FmtEvent<'a> {
-    fn record_debug(&mut self, field: &Field, value: &dyn fmt::Debug) {
-        let buf = &mut self.bufs.current_buf;
-        write!(buf, "{comma} ", comma = if self.comma { "," } else { "" },).unwrap();
-        let name = field.name();
-        if name == "message" {
-            write!(buf, "{:?}", value).unwrap();
-            self.comma = true;
-        } else {
-            write!(buf, "{}={:?}", name, value).unwrap();
-            self.comma = true;
-        }
-    }
+#[derive(Debug)]
+pub struct HierarchicalLayer<W = fn() -> io::Stdout>
+where
+    W: MakeWriter + 'static,
+{
+    make_writer: W,
+    bufs: Mutex<Buffers>,
+    config: Config,
 }
 
-impl<'a> FmtEvent<'a> {
-    fn finish(&mut self, indent: usize, config: &Config) {
-        self.bufs.current_buf.push('\n');
-        self.bufs.indent_current(indent, config);
-        self.bufs.flush_indent_buf();
-    }
-}
-
-struct ColorLevel<'a>(&'a Level);
-
-impl<'a> fmt::Display for ColorLevel<'a> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match *self.0 {
-            Level::TRACE => Color::Purple.bold().paint("TRACE"),
-            Level::DEBUG => Color::Blue.bold().paint("DEBUG"),
-            Level::INFO => Color::Green.bold().paint(" INFO"),
-            Level::WARN => Color::RGB(252, 234, 160).bold().paint(" WARN"), // orange
-            Level::ERROR => Color::Red.bold().paint("ERROR"),
-        }
-        .fmt(f)
-    }
-}
-
-fn indent_block_with_lines(lines: &[&str], buf: &mut String, indent: usize, indent_amount: usize) {
-    let indent_spaces = indent * indent_amount;
-    if lines.len() == 0 {
-        return;
-    } else if indent_spaces == 0 {
-        for line in lines {
-            buf.push_str(line);
-            buf.push('\n');
-        }
-        return;
-    }
-    let mut s = String::with_capacity(indent_spaces);
-
-    // instead of using all spaces to indent, draw a vertical line at every indent level
-    // up until the last indent
-    for i in 0..(indent_spaces - indent_amount) {
-        if i % indent_amount == 0 {
-            s.push_str(LINE_VERT);
-        } else {
-            s.push(' ');
-        }
-    }
-
-    // draw branch
-    buf.push_str(&s);
-    buf.push_str(LINE_BRANCH);
-
-    // add `indent_amount - 1` horizontal lines before the span/event
-    for _ in 0..(indent_amount - 1) {
-        buf.push_str(LINE_HORIZ);
-    }
-    buf.push_str(&lines[0]);
-    buf.push('\n');
-
-    // add the rest of the indentation, since we don't want to draw horizontal lines
-    // for subsequent lines
-    for i in 0..indent_amount {
-        if i % indent_amount == 0 {
-            s.push_str(LINE_VERT);
-        } else {
-            s.push(' ');
-        }
-    }
-
-    // add all of the actual content, with each line preceded by the indent string
-    for line in &lines[1..] {
-        buf.push_str(&s);
-        buf.push_str(line);
-        buf.push('\n');
-    }
-}
-
-fn indent_block(
-    block: &mut String,
-    buf: &mut String,
-    indent: usize,
-    indent_amount: usize,
-    indent_lines: bool,
-) {
-    let lines: Vec<&str> = block.lines().collect();
-    let indent_spaces = indent * indent_amount;
-    buf.reserve(block.len() + (lines.len() * indent_spaces));
-    if indent_lines {
-        indent_block_with_lines(&lines, buf, indent, indent_amount);
-    } else {
-        let indent_str = String::from(" ").repeat(indent_spaces);
-        for line in lines {
-            buf.push_str(&indent_str);
-            buf.push_str(line);
-            buf.push('\n');
-        }
-    }
-}
-
-impl HierarchicalLayer {
-    pub fn new(indent_amount: usize) -> Self {
+impl Default for HierarchicalLayer {
+    fn default() -> Self {
         let ansi = atty::is(atty::Stream::Stdout);
+        let indent_amount = 2;
         let config = Config {
             ansi,
             indent_amount,
             ..Default::default()
         };
         Self {
-            stdout: io::stdout(),
+            make_writer: io::stdout,
             bufs: Mutex::new(Buffers::new()),
             config,
         }
     }
+}
 
+impl<W> HierarchicalLayer<W>
+where
+    W: MakeWriter,
+{
     pub fn with_ansi(self, ansi: bool) -> Self {
         Self {
             config: self.config.with_ansi(ansi),
             ..self
         }
+    }
+
+    pub fn with_writer<W2>(self, make_writer: W2) -> HierarchicalLayer<W2>
+    where
+        W2: MakeWriter + 'static,
+    {
+        HierarchicalLayer {
+            make_writer,
+            config: self.config,
+            bufs: self.bufs,
+        }
+    }
+
+    pub fn with_indent_amount(self, indent_amount: usize) -> Self {
+        let config = Config {
+            indent_amount,
+            ..self.config
+        };
+        Self { config, ..self }
     }
 
     pub fn with_indent_lines(self, indent_lines: bool) -> Self {
@@ -290,9 +136,10 @@ impl HierarchicalLayer {
     }
 }
 
-impl<S> Layer<S> for HierarchicalLayer
+impl<S, W> Layer<S> for HierarchicalLayer<W>
 where
     S: Subscriber + for<'span> LookupSpan<'span> + fmt::Debug,
+    W: MakeWriter + 'static,
 {
     fn new_span(&self, attrs: &Attributes, id: &Id, ctx: Context<S>) {
         let data = Data::new(attrs);
@@ -334,7 +181,8 @@ where
 
         bufs.indent_current(indent, &self.config);
         bufs.flush_indent_buf();
-        bufs.flush_current_buf(self.stdout.lock());
+        let writer = self.make_writer.make_writer();
+        bufs.flush_current_buf(writer)
     }
 
     fn on_event(&self, event: &Event<'_>, ctx: Context<S>) {
@@ -392,7 +240,8 @@ where
         };
         event.record(&mut visitor);
         visitor.finish(indent, &self.config);
-        bufs.flush_current_buf(self.stdout.lock());
+        let writer = self.make_writer.make_writer();
+        bufs.flush_current_buf(writer)
     }
 
     fn on_exit(&self, _id: &Id, _ctx: Context<S>) {}
